🐞 fix & 🔧 build: 更换 ext4_rs 库

+++ os/src/fs/ext4/defs.rs
@@ -1,1 +1,1 @@
-pub const ROOT_INO: u64 = 2;
+pub const ROOT_INO: u32 = 2;
